aspectj/book/aspectj-in-action-code/ch16/workspace/Section16.1.3SimplifyingTheDomainObjectAPI/src/main/java/ajia/util/GenericRepository.java
/*
Copyright 2009 Ramnivas Laddad

Licensed under the Apache License, Version 2.0 (the "License"); 
you may not use this file except in compliance with the License. 
You may obtain a copy of the License at 
    http://www.apache.org/licenses/LICENSE-2.0 

Unless required by applicable law or agreed to in writing, software 
distributed under the License is distributed on an "AS IS" BASIS, 
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. 
See the License for the specific language governing permissions and 
limitations under the License. 
*/

//Listing A.6 Reusable generic-based repository interface
package ajia.util;

import java.util.List;

public interface GenericRepository<T> {
    public List<T> findAll();

    public T find(Long id);

    public void update(T entity);

    public void delete(T entity);
}